DPA-171 materials options done
