+++ src/Components/Settings/scss/_filament-options.scss
@@ -2,7 +2,7 @@
 @import '../../UI/shared/functions';
 
 .filament-options {
-  width: 90vw;
+  width: 85vw;
 
   .minMax-wrapper {
     width: 100%;
