+++ src/endpoints.json
@@ -12,6 +12,9 @@
   "filamentsPage": "/filaments-list",
   "settingsPage": "/settings",
   "settingsFilamentsOptions": "/filaments-options",
+  "settingMaterialsOptions": "/material-options",
+  "settingColorsOptions": "/colors-options",
+  "settingBrandsOptions": "/brands-options",
   "usersPage": "/users",
   "ownersPage": "/owners",
 
@@ -44,11 +47,11 @@
   "settingPasswordChange": "/api/account/settings/password/",
   "settingEmailChange": "/api/account/settings/email/",
   "settingFilamentMaterialAdd": "/api/filaments/material/add/",
-  "settingFilamentMaterialDelete": "/-/",
+  "settingFilamentMaterialDelete_id": "/api/filaments/material/delete/",
   "settingFilamentColorAdd": "/api/filaments/color/add/",
-  "settingFilamentColorDelete": "/-/",
+  "settingFilamentColorDelete_id": "//api/filaments/color/delete/",
   "settingFilamentBrandAdd": "/api/filaments/brand/add/",
-  "settingFilamentBrandDelete": "/-/",
+  "settingFilamentBrandDelete_id": "/api/filaments/brand/delete/",
 
   "usersGetAll": "/api/companies/users/get/",
   "userGetOne_id": "/api/companies/users/get/",
